Refactor launcher: isolate desktop logic into launcher.py, clean app.py/requirements, update build scripts, and add tests
